feat(collections): stabilize iterable input for slidingWindows
